requirements.txt
# 合併依賴
pybaseball>=2.2.7
pandas>=2.0.0
numpy
scipy
google-generativeai>=0.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
plotly>=5.18.0
